pyproject.toml
[tool.poetry]
name = "whylogs"
version = "0.6.3-dev0"
description = "Profile and monitor your ML data pipeline end-to-end"
authors = ["WhyLabs.ai <support@whylabs.ai>"]
license = "Apache-2.0"
include = [
  "src/whylogs/proto/*.py",
  "README.md",
  "CHANGELOG.rst",
  "CODE_OF_CONDUCT.md",
  "DEVELOPMENT.md"
]
readme = "README.md"
repository = "https://github.com/whylabs/whylogs"
homepage = "https://docs.whylabs.ai"
keywords = ["machine learning", "data profiling", "data quality", "data logging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apple Public Source License",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = ">=3.6.1,<3.10"
click = ">=7.1.2"
python-dateutil = ">=2.8.1"
protobuf = ">=3.15.5"
pyyaml = ">=5.3.1"
pandas = ">=1.0.0"
marshmallow = ">=3.7.1"
numpy = ">=1.18.0"
whylabs-client = "@>=0.1"
whylabs-datasketches = ">=2.2.0b1"
boto3 = ">=1.14.1"
botocore = ">=1.17.44"
smart-open = ">=4.1.2"
xlrd = "^2.0.1"
puremagic = "@1.10"
matplotlib = "^3.0.3"
mlflow = "~1.13."
tqdm = "^4.60.0"
openpyxl = "^3.0.7"
scikit-learn = "^0.24.2"

[tool.poetry.dev-dependencies]
pytest = "6.1.2"
Pillow = "^8.2.0"
pytest-cov = "^2.11.1"
freezegun = "^1.1.0"
moto = "^1.3.16"
bump2version = "^1.0.1"
colorama = "^0.4.4"
black = "^20.8b1"
nbconvert = "^6.0.7"
ipykernel = "^5.5.3"
setuptools = "^56.0.0"
2to3 = "^1.0"
sphinx-rtd-theme = "^0.5.2"
autoapi = "^2.0.1"
sphinx-autoapi = "^1.8.0"
sphinx-autodoc-typehints = "^1.12.0"
watchdog = { extras = ["watchmedo"], version = "^2.0.2" }
isort = "^5.8.0"
autoflake = "^1.4"
flake8 = "^3.9.1"
aiohttp = "^3.7.4"
aiohttp_cors = "^0.7.0"
ipywidgets = "^7.6.3"
jiwer = "^2.2.0"
pyarrow = "^5.0.0"
pybars3 = "^0.9.7"
twine = "^3.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
skip = ["src/whylogs/proto/", ".venv"]

[tool.black]
line-length = 160
target-version = ["py38"]
include_trailing_comma = "True"
exclude = "src/whylogs/proto/"
force_grid_wrap = 0
use_parentheses = "True"
ensure_newline_before_comments = "True"
